refactor: replace C-style casts with static_cast/const_cast
- enable google-readability-casting check
- replace 64 C-style casts across 19 source files
- remove unnecessary const_cast in KeyFrameDatabase::SetORBVocabulary
- add TODO for removing const_cast in SerializationUtils load path

+++ src/Frame.cc
@@ -1082,7 +1082,7 @@ void Frame::ComputeStereoMatches() {
       }
 
       // Re-scaled coordinate
-      float bestuR = mvScaleFactors[kpL.octave] * ((float)scaleduR0 + (float)bestincR + deltaR);
+      float bestuR = mvScaleFactors[kpL.octave] * (scaleduR0 + static_cast<float>(bestincR) + deltaR);
 
       float disparity = (uL - bestuR);
